backend/requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-rest-framework-simplejwt-blacklist==0.1.1
Pillow==10.0.1
python-decouple==3.8
PyJWT==2.8.0
cryptography==41.0.7
